refactor to reduce verbosity of request seraializers
