+++ rust/src/endpoints/get_leader.rs
@@ -29,6 +29,21 @@ const GET_LEADER_REPLY_FILE_IDENTIFIER: ParsedFileIdentifier = ParsedFileIdentif
     file_identifier_name: Some("LeaderInfo"),
 };
 
+// fn serialize_request(peer: SocketAddr, cluster_key: &str, known_leader: UID) -> Result<FlowMessage> {
+//     // key, known_leader, reply_promise, directly in fake_root
+//     use crate::common_generated::{ReplyPromise, ReplyPromiseArgs};
+//     use get_leader_request::{FakeRoot, FakeRootArgs};
+
+//     let completion = UID::random_token();
+//     let wltoken = UID::well_known_token(WLTOKEN::ClientLeaderRegGetLeader);
+
+//     let mut builder = flatbuffers::FlatBufferBuilder::with_capacity(1024);
+//     let response_token: crate::common_generatedUID = (&completion).into();
+//     let uid = Some(&response_token);
+//     let reply_promise = Some(ReplyPromise::create())
+
+// }
+
 fn deserialize_request(buf: &[u8]) -> Result<GetLeaderRequest> {
     let fake_root = get_leader_request::root_as_fake_root(buf)?;
     let request = fake_root.request().unwrap();
@@ -63,7 +78,6 @@ fn test_leader_info() -> Result<()> {
         0x00, 0x00, 0x1e, 0x0e, 0xe3, 0xd1, 0x00, 0x00, 0x00, 0x00, 0x07, 0x00, 0x00, 0x00, 0x74,
         0x65, 0x73, 0x74, 0x4b, 0x65, 0x79, 0x00,
     ];
-    // let buffer_10LeaderInfo = vec![0x1c, 0x00, 0x00, 0x00, 0x6a, 0x3d, 0x7f, 0x00, 0x00, 0x00, 0x00, 0x00, 0x06, 0x00, 0x08, 0x00, 0x04, 0x00, 0x0a, 0x00, 0x19, 0x00, 0x04, 0x00, 0x14, 0x00, 0x18, 0x00, 0x10, 0x00, 0x00, 0x00, 0x04, 0x00, 0x00, 0x00, 0x12, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x0c, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00];
     let buffer_7ErrorOrI10LeaderInfoE = vec![
         0x28, 0x00, 0x00, 0x00, 0x6a, 0x3d, 0x7f, 0x02, 0x00, 0x00, 0x06, 0x00, 0x08, 0x00, 0x04,
         0x00, 0x06, 0x00, 0x06, 0x00, 0x04, 0x00, 0x08, 0x00, 0x09, 0x00, 0x08, 0x00, 0x04, 0x00,
